Save "Swap time and date positions" checkbox state
Currently, if the user checks the option "Swap time and date positions",
leaves the settings modal and returns to it, the option will be
unchecked, even though the date and time are, indeed, swapped.

This saves the value of the "Swap time and date positions" option so
that it will match whatever state the user selected last.

Fixes: #165