Increase thread message limit
I've had no issues going as high as 40 messages every few seconds. May make the loading time longer, though. 

I think a slider in the settings to set the thread and thread message limit would be ideal, since I rarely want to load any more than the top 2-3 threads, but it would be better to load more messages in those threads. But on my device, 10 is too low and a lot of time is spent loading old messages.