Update strings.xml
Cleaned up the wording in the Backup and Restore section to make it clearer and more grammatically correct.

The only issue is I wonder what will happen because I extended the length of string by 2 characters for the password check mark option on line 16. Will it wrap or make the dialog bigger?